changed main.rs dimesions constants types to more ideal

+++ life/src/main.rs
@@ -7,13 +7,13 @@ use sdl2::event::Event;
 use sdl2::keyboard::Keycode;
 use sdl2::pixels;
 
-const WIDTH: i32 = 400;
-const HEIGHT: i32 = 400;
+const WIDTH: u32 = 400;
+const HEIGHT: u32 = 400;
 
 fn main() {
     let sdl_context = sdl2::init().unwrap();
     let video = sdl_context.video().unwrap();
-    let window = video.window("GOL", WIDTH as u32, HEIGHT as u32)
+    let window = video.window("GOL", WIDTH, HEIGHT)
         .position_centered().build().unwrap();
     let mut canvas = window.into_canvas().build().unwrap();
     let mut events = sdl_context.event_pump().unwrap();
